scripts v2.0.1: only change FnMODE if user input is valid
Calling XCMD("1 FnMODE $FM") leads to a TopSpin error "illegal
enumeration type" if $FM is not a valid value for FnMODE.

In particular, this can happen if the noah_xxxx AU programme is called
via `xaup`, which (inexplicably) appends "xaup" to the argument list.
Thus, when `xaup` was used to run noah_xxxx without any other arguments,
argv[2] would be "xaup" which would then trigger the "illegal
enumeration" error.

We therefore skip setting FnMODE if argv[2] is "xaup", and on invalid
input display a non-blocking dialog box to alert the user.