website, pprogs: add F2-coupled HSQC modules (Sj)
Website headings have now been renamed to a less suggestive "13C-1H #1"
and "13C-1H #2", because in general we can have a number of modules in
the first slot.

However, the interface is getting a bit clunky with repeated
information. It seems that eventually we may need to redesign it.