website, pprogs: rename modules, introduce [ID] substitution
The module renaming was mainly done in order to make sure that newly
defined delays did not exceed TopSpin's length limits (15 chars).

However, I also did not like the fact that the delay names did not
always correspond to the module code, so I *enforced* this by using [ID]
in the delay names (and gradient list names), This breaks some backwards
compatibility, but as far as I'm aware I'm the only one who knows about
the devmode modules, so we may as well break it in a glorious fashion by
renaming a ton of modules at once.