pprogs: revert double ES in CX modules
It doesn't behave as expected. :-( Adding ES inside the COSY part of CX
modules leads to 'bonus suppression' of peaks in the X module, at
frequencies which correspond to nulls in the inversion profile of the
soft 180 pulse used in the ES block.

Probably we will have to make do with presat for these cases.